moved folder into api folder, added error handling, remastered folder move
