+++ package.json
@@ -18,6 +18,7 @@
     "mongoose": "^6.3.0",
     "morgan": "^1.10.0",
     "multer": "^1.4.4",
+    "nodemon": "^2.0.15",
     "react": "^18.0.0",
     "react-dom": "^18.0.0",
     "react-scripts": "5.0.0",
