+++ src/api/routes/users.js
@@ -2,7 +2,7 @@ const bodyParser = require('body-parser');
 const express = require('express');
 const router = express.Router();
 const jsonParser = bodyParser.json();
-const checkAdmin = require('./../../middleware/check-admin-auth');
+const checkAdmin = require('./../middleware/check-admin-auth');
 const uuid = require('uuid');
 const bcrypt = require('bcrypt');
 const jwt = require('jsonwebtoken')
