+++ src/server.js
@@ -4,7 +4,7 @@ const mongoose = require('mongoose');
 const morgan = require('morgan');
 
 const multer = require('multer');
-const cors = require('./middleware/cors');
+const cors = require('./api/middleware/cors');
 
 const http = require('http');
 const app = express();
@@ -85,6 +85,21 @@ app.use('/users', userRoutes);
 app.use(cors);
 app.use(express.static("public"));
 app.use(morgan('dev'));
+
+app.use((req, res, next) => {
+    const error = new Error("not found");
+    error.status = 404;
+    next(error);
+});
+
+app.use((error, req, res, next) => {
+    res.status(error.status || 500 );
+    res.json({
+        error:{
+            message:error.message
+        }
+    });
+})
 
 ////------------------>SERVER<------------------
 app.listen(PORT, () => {
